requirements.txt
jupyter
numpy
pandas
tensorflow-gpu==2.8.0
scikit-learn
pathlib
matplotlib